Revert "API updated."
This reverts commit ab95c1762b19257ae77cf90185be00dfdf9a061a.

+++ DJWatermelon/AudioService/Lavalink/Models/PlayerState.cs
@@ -12,10 +12,10 @@ namespace DJWatermelon.AudioService.Lavalink.Models;
 public sealed record class PlayerState(
     [property: JsonRequired]
     [property: JsonPropertyName("time")]
-    ulong AbsoluteTimestamp,
+    int AbsoluteTimestamp,
 
     [property: JsonPropertyName("position")]
-    ulong Position,
+    int Position,
 
     [property: JsonRequired]
     [property: JsonPropertyName("connected")]
